// workspaces/mooncg/src/types/mooncg.ts
type Person =
	| {
			name: string;
			email?: string;
			url?: string;
	  }
	| string;

export namespace MoonCG {
	/**
	 * A collection of types that represent the raw data from the `mooncg` stanza in a bundle's `package.json`.
	 */
	export namespace Manifest {
		/**
		 * An unparsed `assetCategory` from a bundle's `package.json`.
		 */
		export interface UnparsedAssetCategory {
			name: string;
			title: string;
			allowedTypes?: string[];
		}

		/**
		 * An unparsed `panel` from a bundle's `package.json`.
		 */
		export interface UnparsedPanel {
			name: string;
			title: string;
			file: string;
			headerColor?: string;
			fullbleed?: boolean;
			workspace?: string;
			dialog?: boolean;
			dialogButtons?: { name: string; type: "dismiss" | "confirm" }[];
			width?: number;
		}

		/**
		 * An unparsed `graphic` from a bundle's `package.json`.
		 */
		export interface UnparsedGraphic {
			file: string;
			width: number;
			height: number;
			singleInstance?: boolean;
		}

		/**
		 * An unparsed `mount` configuration from a bundle's `package.json`.
		 */
		export interface UnparsedMount {
			directory: string;
			endpoint: string;
		}

		/**
		 * An unparsed `soundCue` from a bundle's `package.json`.
		 */
		export interface UnparsedSoundCue {
			name: string;
			assignable?: boolean;
			defaultVolume?: number;
			defaultFile?: string;
		}

		/**
		 * An unparsed record of `bundleDependencies` from a bundle's `package.json`.
		 */
		export type UnparsedBundleDependencies = Record<string, string>;

		/**
		 * The actual structure of a bundle's `mooncg` stanza from it's `package.json`.
		 */
		export interface UnparsedManifest {
			compatibleRange?: string;
			transformBareModuleSpecifiers?: boolean;
			dashboardPanels?: UnparsedPanel[];
			graphics?: UnparsedGraphic[];
			assetCategories?: UnparsedAssetCategory[];
			mount?: UnparsedMount[];
			soundCues?: UnparsedSoundCue[];
			bundleDependencies?: UnparsedBundleDependencies;
		}
	}

	/**
	 * The rough structure of a bundle's `package.json`.
	 * Only includes types that MoonCG cares about.
	 */
	export interface PackageJSON {
		name: string;
		version: string;
		license?: string;
		description?: string;
		homepage?: string;
		author?: Person;
		contributors?: Person[];
		mooncg?: Manifest.UnparsedManifest;
	}

	/**
	 * A _parsed_ bundle manifest as used internally by MoonCG.
	 */
	export type Manifest = Omit<PackageJSON, "mooncg"> &
		Manifest.UnparsedManifest & { transformBareModuleSpecifiers: boolean };
// ...
}